Make all LinphoneGenericActivity activities fullscreen
