+++ app/src/main/java/org/linphone/activities/LinphoneGenericActivity.java
@@ -22,6 +22,7 @@ package org.linphone.activities;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.Surface;
+import android.view.View;
 import org.linphone.LinphoneContext;
 import org.linphone.LinphoneManager;
 import org.linphone.core.Core;
@@ -39,6 +40,8 @@ public abstract class LinphoneGenericActivity extends ThemeableActivity {
     @Override
     protected void onResume() {
         super.onResume();
+
+        makeFullscreen();
 
         ensureServiceIsRunning();
 
@@ -73,6 +76,17 @@ public abstract class LinphoneGenericActivity extends ThemeableActivity {
                 core.setDeviceRotation(rotation);
             }
         }
+    }
+
+    public void makeFullscreen() {
+        View decorView = getWindow().getDecorView();
+        decorView.setSystemUiVisibility(
+                View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
+                        | View.SYSTEM_UI_FLAG_LAYOUT_STABLE
+                        | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
+                        | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
+                        | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
+                        | View.SYSTEM_UI_FLAG_FULLSCREEN);
     }
 
     private void ensureServiceIsRunning() {
