+++ app/src/main/java/org/linphone/activities/MainActivity.java
@@ -658,15 +658,4 @@ public abstract class MainActivity extends LinphoneGenericActivity
     public SideMenuFragment getSideMenuFragment() {
         return mSideMenuFragment;
     }
-
-    public void makeFullscreen() {
-        View decorView = getWindow().getDecorView();
-        decorView.setSystemUiVisibility(
-                View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
-                        | View.SYSTEM_UI_FLAG_LAYOUT_STABLE
-                        | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
-                        | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-                        | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
-                        | View.SYSTEM_UI_FLAG_FULLSCREEN);
-    }
 }
